binfmt: fix build warnings
binfmt_execmodule.c:181:29: warning: passing argument 2 of 'mm_initialize' makes pointer from integer without a cast [-Wint-conversion]
  mm_initialize(binp->uheap, binp->heapstart + sizeof(struct mm_heap_s), binp->uheap_size);
                             ^~~~
binfmt_execmodule.c:197:2: warning: implicit declaration of function 'mpu_get_register_value' [-Wimplicit-function-declaration]
  mpu_get_register_value(&rtcb->mpu_regs[3], g_mpu_region_nr, (uintptr_t)binp->alloc[ALLOC_TEXT], binp->textsize, true, true);
  ^~~~~~~~~~~~~~~~~~~~~~
binfmt_execmodule.c:345:17: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
  tcb->cmn.uheap = binp->uheap;
                 ^
binfmt_exec.c:148:16: warning: unused variable 'tcb' [-Wunused-variable]
  struct tcb_s *tcb;
                ^~~

binfmt_loadbinary.c:118:10: warning: passing argument 1 of 'memcpy' makes pointer from integer without a cast [-Wint-conversion]
   memcpy(bin->datastart, bin->data_backup, bin->datasize);
          ^~~
binfmt_loadbinary.c:118:26: warning: passing argument 2 of 'memcpy' makes pointer from integer without a cast [-Wint-conversion]
   memcpy(bin->datastart, bin->data_backup, bin->datasize);
                          ^~~
binfmt_loadbinary.c:119:10: warning: passing argument 1 of 'memset' makes pointer from integer without a cast [-Wint-conversion]
   memset(bin->bssstart, 0, bin->bsssize);
          ^~~
binfmt_loadbinary.c:171:52: warning: passing argument 3 of 'mpu_get_register_value' makes integer from pointer without a cast [-Wint-conversion]
    mpu_get_register_value(NULL, g_mpu_region_nr++, bin->alloc[ALLOC_TEXT], bin->textsize, true, true);
                                                    ^~~
binfmt_loadbinary.c:172:52: warning: passing argument 3 of 'mpu_get_register_value' makes integer from pointer without a cast [-Wint-conversion]
    mpu_get_register_value(NULL, g_mpu_region_nr++, bin->alloc[ALLOC_RO], bin->rosize, true, false);
                                                    ^~~
binfmt_loadbinary.c:173:52: warning: passing argument 3 of 'mpu_get_register_value' makes integer from pointer without a cast [-Wint-conversion]
    mpu_get_register_value(NULL, g_mpu_region_nr++, bin->alloc[ALLOC_DATA], bin->ramsize, false, false);
                                                    ^~~
binfmt_loadbinary.c:188:10: warning: passing argument 1 of 'memcpy' makes pointer from integer without a cast [-Wint-conversion]
   memcpy(bin->data_backup, bin->datastart, bin->datasize);
          ^~~
binfmt_loadbinary.c:188:28: warning: passing argument 2 of 'memcpy' makes pointer from integer without a cast [-Wint-conversion]
   memcpy(bin->data_backup, bin->datastart, bin->datasize);
                            ^~~
binfmt_loadbinary.c:199:26: warning: initialization makes pointer from integer without a cast [-Wint-conversion]
   uint32_t *heap_table = (uint32_t)g_lib_binp->alloc[ALLOC_DATA] + 4;
                          ^

libelf/libelf_bind.c:201:9: warning: assignment makes pointer from integer without a cast [-Wint-conversion]
    prel = loadinfo->reltab + sizeof(Elf32_Rel) * i;
         ^
libelf/libelf_bind.c:224:9: warning: assignment makes pointer from integer without a cast [-Wint-conversion]
    psym = loadinfo->symtab + sizeof(Elf32_Sym) * symidx;
         ^
libelf/libelf_bind.c:274:11: warning: passing argument 1 of 'kmm_free' makes pointer from integer without a cast [-Wint-conversion]
  kmm_free(loadinfo->reltab);
           ^~~~~~~~
libelf/libelf_bind.c:275:19: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
  loadinfo->reltab = NULL;
                   ^
libelf/libelf_bind.c:310:9: warning: assignment makes pointer from integer without a cast [-Wint-conversion]
    psym = loadinfo->symtab + sizeof(Elf32_Sym) * i;
         ^
libelf/libelf_bind.c:320:10: warning: implicit declaration of function 'elf_symname' [-Wimplicit-function-declaration]
    ret = elf_symname(loadinfo, psym);
          ^~~~~~~~~~~
libelf/libelf_bind.c:393:2: warning: implicit declaration of function 'elf_readstrtab' [-Wimplicit-function-declaration]
  elf_readstrtab(loadinfo);
  ^~~~~~~~~~~~~~
libelf/libelf_bind.c:485:12: warning: passing argument 1 of 'kmm_free' makes pointer from integer without a cast [-Wint-conversion]
   kmm_free(loadinfo->strtab);
            ^~~~~~~~
libelf/libelf_bind.c:486:20: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
   loadinfo->strtab = NULL;
                    ^
libelf/libelf_bind.c:488:11: warning: passing argument 1 of 'kmm_free' makes pointer from integer without a cast [-Wint-conversion]
  kmm_free(loadinfo->symtab);
           ^~~~~~~~
libelf/libelf_bind.c:489:19: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
  loadinfo->symtab = NULL;
                   ^

libelf/libelf_addrenv.c:171:22: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
  loadinfo->textalloc = kumm_memalign(loadinfo->textsize, loadinfo->textsize);
                      ^
libelf/libelf_addrenv.c:172:20: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
  loadinfo->roalloc = kumm_memalign(loadinfo->rosize, loadinfo->rosize);
                    ^
libelf/libelf_addrenv.c:173:22: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
  loadinfo->dataalloc = kumm_memalign(datamemsize, datamemsize);
                      ^

libelf/libelf_load.c:247:29: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
    loadinfo->binp->bssstart = *pptr;
                             ^

libelf/libelf_symbols.c:108:27: warning: passing argument 2 of 'elf_read' makes pointer from integer without a cast [-Wint-conversion]
  ret = elf_read(loadinfo, loadinfo->strtab, strtab->sh_size, strtab->sh_offset);
                           ^~~~~~~~
libelf/libelf_symbols.c:111:12: warning: passing argument 1 of 'kmm_free' makes pointer from integer without a cast [-Wint-conversion]
   kmm_free(loadinfo->strtab);
            ^~~~~~~~
libelf/libelf_symbols.c:112:20: warning: assignment makes integer from pointer without a cast [-Wint-conversion]
   loadinfo->strtab = NULL;
                    ^
libelf/libelf_symbols.c:162:21: warning: assignment makes pointer from integer without a cast [-Wint-conversion]
  loadinfo->iobuffer = loadinfo->strtab + sym->st_name;
                     ^
libelf/libelf_symbols.c:224:25: warning: passing argument 2 of 'elf_read' makes pointer from integer without a cast [-Wint-conversion]
  if (elf_read(loadinfo, loadinfo->symtab, symtab->sh_size, symtab->sh_offset) < 0) {
                         ^~~~~~~~
libelf/libelf_symbols.c: In function 'elf_symvalue':
libelf/libelf_symbols.c:331:29: warning: implicit declaration of function 'hashmap_get' [-Wimplicit-function-declaration]
   sym->st_value = (uint32_t)hashmap_get((struct hashmap_s *) exports, hashmap_get_hashval(loadinfo->iobuffer));
                             ^~~~~~~~~~~
libelf/libelf_symbols.c:331:71: warning: implicit declaration of function 'hashmap_get_hashval' [-Wimplicit-function-declaration]
   sym->st_value = (uint32_t)hashmap_get((struct hashmap_s *) exports, hashmap_get_hashval(loadinfo->iobuffer));
                                                                       ^~~~~~~~~~~~~~~~~~~
libelf/libelf_symbols.c:290:29: warning: unused variable 'symbol' [-Wunused-variable]
  FAR const struct symtab_s *symbol;
                             ^~~~~~

+++ os/binfmt/libelf/libelf_load.c
@@ -244,7 +244,7 @@ static inline int elf_loadfile(FAR struct elf_loadinfo_s *loadinfo)
 
 		else {
 #ifdef CONFIG_OPTIMIZE_APP_RELOAD_TIME
-			loadinfo->binp->bssstart = *pptr;
+			loadinfo->binp->bssstart = (uint32_t)*pptr;
 #endif
 			memset(*pptr, 0, shdr->sh_size);
 		}
@@ -253,7 +253,7 @@ static inline int elf_loadfile(FAR struct elf_loadinfo_s *loadinfo)
 
 		binfo("%d. %08lx->%08lx\n", i, (unsigned long)shdr->sh_addr, (unsigned long)*pptr);
 
-		shdr->sh_addr = (uintptr_t) * pptr;
+		shdr->sh_addr = (uintptr_t)*pptr;
 
 		/* Setup the memory pointer for the next time through the loop */
